Split interface and class into separate files
